Implemented JSON -> database importer, retroactively adjusted migrations

+++ QuartierLatin.Backend/Application/Catalog/CommonTraitAppService.cs
@@ -21,7 +21,7 @@ namespace QuartierLatin.Backend.Application.Catalog
             return await _commonTraitRepository.GetCommonTraitListByTypeId(typeId);
         }
 
-        public async Task<int> CreateCommonTraitAsync(int typeId, JObject names, int order, long? iconBlobId, int? parentId)
+        public async Task<int> CreateCommonTraitAsync(int typeId, JObject names, int order, int? iconBlobId, int? parentId)
         {
             return await _commonTraitRepository.CreateCommonTraitAsync(typeId, names, iconBlobId, order, parentId);
         }
@@ -31,7 +31,7 @@ namespace QuartierLatin.Backend.Application.Catalog
             return await _commonTraitRepository.GetCommonTraitAsync(id);
         }
 
-        public async Task UpdateCommonTraitAsync(int id, JObject names, int commonTraitTypeId, long? iconBlobId,
+        public async Task UpdateCommonTraitAsync(int id, JObject names, int commonTraitTypeId, int? iconBlobId,
             int order, int? parentId)
         {
             await _commonTraitRepository.UpdateCommonTraitAsync(id, commonTraitTypeId, names, iconBlobId, order, parentId);
