+++ QuartierLatin.Backend/Application/Interfaces/Catalog/ICommonTraitAppService.cs
@@ -8,8 +8,8 @@ namespace QuartierLatin.Backend.Application.Interfaces.Catalog
     public interface ICommonTraitAppService
     {
         Task<List<CommonTrait>> GetTraitOfTypesByTypeIdAsync(int typeId);
-        Task<int> CreateCommonTraitAsync(int typeId, JObject names, int order, long? iconBlobId, int? parentId);
+        Task<int> CreateCommonTraitAsync(int typeId, JObject names, int order, int? iconBlobId, int? parentId);
         Task<CommonTrait> GetTraitByIdAsync(int id);
-        Task UpdateCommonTraitAsync(int id, JObject names, int commonTraitTypeId, long? iconBlobId, int order, int? parentId);
+        Task UpdateCommonTraitAsync(int id, JObject names, int commonTraitTypeId, int? iconBlobId, int order, int? parentId);
     }
 }
