+++ QuartierLatin.Backend/Database/Migrations/CommonTraits/CreateCommonTraitsToUniversities.cs
@@ -8,8 +8,8 @@ namespace QuartierLatin.Backend.Database.Migrations.CommonTraits
         public override void Up()
         {
             Create.Table("CommonTraitsToUniversities")
-                .WithColumn("UniversityId").AsInt64().ForeignKey("Universities", "Id").PrimaryKey()
-                .WithColumn("CommonTraitId").AsInt64().ForeignKey("CommonTraits", "Id").PrimaryKey();
+                .WithColumn("UniversityId").AsInt32().ForeignKey("Universities", "Id").PrimaryKey()
+                .WithColumn("CommonTraitId").AsInt32().ForeignKey("CommonTraits", "Id").PrimaryKey();
         }
     }
 }
