+++ QuartierLatin.Backend/Storages/IBlobFileStorage.cs
@@ -5,10 +5,10 @@ namespace QuartierLatin.Backend.Storages
 {
     public interface IBlobFileStorage
     {
-        Task CreateBlobAsync(long id, Stream s, int? dimension = null);
-        Stream OpenBlob(long id, int? dimension = null);
-        Task DeleteBlob(long id, int? dimension = null);
+        Task CreateBlobAsync(int id, Stream s, int? dimension = null);
+        Stream OpenBlob(int id, int? dimension = null);
+        Task DeleteBlob(int id, int? dimension = null);
 
-        bool CheckIfExist(long id, int? dimension = null);
+        bool CheckIfExist(int id, int? dimension = null);
     }
 }
